Stok Kısmında görüntüde güncellemeler

+++ Restoran Otomasyon/Data/Kampanya.cs
@@ -9,7 +9,7 @@ namespace Restoran_Otomasyon.Data
 
 		public string Kod { get; set; }
 
-		public int Indirim { get; set; }
+		public decimal Indirim { get; set; }
 
 		public DateTime GecerlilikTarihi { get; set; }
