+++ Restoran Otomasyon/Data/Menu.cs
@@ -11,7 +11,7 @@ namespace Restoran_Otomasyon.Data
 
 		public string Aciklama { get; set; }
 
-		public int Fiyat { get; set; }
+		public decimal Fiyat { get; set; }
 
 		public string Detay { get; set; }
