+++ Restoran Otomasyon/Data/Siparis.cs
@@ -11,7 +11,7 @@ namespace Restoran_Otomasyon.Data
 
 		public string Adres { get; set; }
 
-		public int Tutar { get; set; }
+		public decimal Tutar { get; set; }
 
 		public bool OdemeDurum { get; set; }
